split filer read into 2 files
